+++ packages/framer-motion/src/utils/process.ts
@@ -1,8 +1,0 @@
-/**
- * Browser-safe usage of process
- */
-const defaultEnvironment = "production"
-export const env =
-    typeof process === "undefined" || process.env === undefined
-        ? defaultEnvironment
-        : process.env.NODE_ENV || defaultEnvironment