+++ packages/framer-motion/src/value/scroll/use-viewport-scroll.ts
@@ -5,6 +5,11 @@ import { useScroll } from "../use-scroll"
  * @deprecated useViewportScroll is deprecated. Convert to useScroll()
  */
 export function useViewportScroll() {
-    warnOnce(false, "useViewportScroll is deprecated. Convert to useScroll().")
+    if (process.env.NODE_ENV !== "production") {
+        warnOnce(
+            false,
+            "useViewportScroll is deprecated. Convert to useScroll()."
+        )
+    }
     return useScroll()
 }